+++ src/array/flatten.js
@@ -1,0 +1,6 @@
+export function flatten(arr, depth = 1) {
+  return depth > 0
+    ? arr.reduce((acc, val) =>
+        acc.concat(Array.isArray(val) ? flatten(val, depth - 1) : val), [])
+    : arr.slice();
+}